1. create config_cli.cpp and .h file to separate the core design of the cli module and instance creation.
