+++ source/application/STM32F439ZI_LwIP_TCPIP/config/config_cli.h
@@ -1,0 +1,12 @@
+
+#pragma once
+
+#if defined (__cplusplus)
+extern "C" {
+#endif
+
+void putCharIntoBuffer( char c );
+
+#if defined (__cplusplus)
+}
+#endif
